Move the exit full-screen button to the left.
Having the exit full-screen button in the top right corner is dangerous.
Given bad timing, when a presenter closes its screen-sharing session and
a viewer tries to minimize the presenter's shared screen, the viewer may
click the "leave" button instead and leave the conference.

By keeping the minimize button on the left in the same scenario the
viewer will open the chat instead which is still not ideal but more
acceptable.