Changed "mutato" (which means "changed", not "muted") to "silenziato"
In Italia, "mutato" is an Anglicism - a word formed by taking an English word ("muted") and adapting it to Italian. This can be understandable in some cases, when an Italian equivalent of an English term does not exist; it is less understandable when an Italian word does actually exist, as is the case with "mutato": the correct Italian word for it is "silenziato" (which has the same root as "silence", of course). Moreover, "mutato" in Italian actually means a different thing altogether, it means "changed" or "trans**muted**" (here again, you can see the same root in Italian and English, which IIUIC [comes from Latin](https://en.wiktionary.org/wiki/muto#Latin)).

Finally, you can see some examples of "muted" translated to "silenziato" on [reverso](https://context.reverso.net/translation/italian-english/silenziato).

tl;dr = The Italian for "muted" is "silenziato", not "mutato" (which means "changed")